requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
cycler==0.10.0
decorator==4.4.2
Django==3.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
image==1.5.32
imageio==2.9.0
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
networkx==2.4
numpy==1.19.1
opencv-python==4.4.0.40
opt-einsum==3.3.0
pandas==1.1.0
Pillow==7.2.0
protobuf==3.12.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tifffile==2020.7.24
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
